requirements.txt
django
groq
python-dotenv